Extract `Transition` into module and move strategies around
