+++ rust/connlib/tunnel/src/tests/assertions.rs
@@ -1,5 +1,5 @@
-use super::{ReferenceState, TunnelTest};
-use crate::tests::ResourceDst;
+use super::{reference::ReferenceState, TunnelTest};
+use crate::tests::transition::ResourceDst;
 use connlib_shared::DomainName;
 use ip_packet::IpPacket;
 use pretty_assertions::assert_eq;
